Revert "Revert "Luni URLConnectionTest: Don't depend on TLSv1*""
Cherry-pick note: This went to tm-dev but didn't automerge to tm-wear-kr1-dev and downstream (I suspect they weren't in the automerger config at that point).  It's a prereq for landing aosp/2796533 in CTS 12, which is needed for a partner fix.

This reverts commit ad855c0c58d77e8c3c16c6e1f26e317beb2cdc91.

Bug: 288058920
Bug: 292196077

Reason for revert: Fixes the previous failure
Merged-In: Iaaedb093f414546b681172cb35d91e26bfab0e43
Change-Id: Iaaedb093f414546b681172cb35d91e26bfab0e43