Preload HttpEngine for Android versions [S -> V]
HttpEngine is currently Android's most performant networking stack.
However, it's suffering from performance issues during its startup that
are documented in b/380349437 and in order to overcome that, we figured
out that preloading HttpEngine is the best path forward.

We're currently undergoing rolling preloading for Android B+ via an
experiment flag that have shown 20% improvement in P50 and 40%
improvement in the P90 for startup metrics without any regressions in
system health metrics (eg: boot time, memory, startup).

The reason why the preloading is happening in ZygoteHooks is because
it's the only mainline module that is being called from Zygote since
Android S, which means that this is the only place that will be capable
of inserting new code into Zygote. This is a hacky workaround but there
is no other way of achieving this unfortunately.

Test: Flashed Android 14 and installed both Tethering + ART mainline
module built from tip of tree with the change, verified that device
boots and httpEngine is preloaded.
Bug: 380349437

Change-Id: Ic3976669e0eb08a37f7c92e421b61e92b576b0ce