Add API for starting tracing for long running methods.
This adds a new API (startLowOverheadTraceForLongMethods) for starting
a trace that records methods that are executing for longer than a set
threshold. The API also specifies the duration of the trace. If the
trace is stopped (if still running) after the specified duration. This
also renames the existing startLowOverheadTrace to
startLowOverheadTraceForAllMethods to be clear about what is being
traced. The tracing can be stopped using the existing
StopLowOverheadTracing API.

Bug: 352518093
Test: art/test.py
Change-Id: I7561670f78047a0fe02bcd0e33f34c376b7905ad