Override api_surface for intra core
Preparation for	changes	that add configuration for the API surfaces in
`build/soong/java/metalava/main-config.xml` and pass Metalava the name
of the surface instead of `--show-single-annotation` command line
options. That will avoid duplicating the command line arguments for
each API surface in multiple locations.

While most `java_sdk_library` modules use the correct `api_surface`
that is not true for the `intra-core` API so this change overrides the
default `api_surface` for that.

Bug: 391554590
Test: m checkapi
Change-Id: I42e9c69af7152de374f628e21b535476635fb6a3